feat: move components into modules
