+++ preview_generator/src/components.rs
@@ -1,0 +1,7 @@
+mod circle_rect;
+mod rounded_rect;
+mod tags;
+
+pub use circle_rect::*;
+pub use rounded_rect::*;
+pub use tags::*;
